lastPageNumber -> lastProblemId로 변경된 내용 반영

+++ s-with-me-back/src/main/java/com/swithme/web/dto/MyBookResponseDto.java
@@ -11,7 +11,7 @@ public class MyBookResponseDto {
     private int myBookId;
     private int folderId;
     private int bookId;
-    private short lastProblemId;
+    private int lastProblemId;
 
     @Builder
     public MyBookResponseDto(int myBookId, int folderId, int bookId, short lastProblemId){
